Implement some tests cases
