+++ .travis.yml
@@ -1,0 +1,20 @@
+---
+language: nix
+env:
+  global:
+    - STORE=$HOME/nix-store
+cache:
+  directories:
+    - $STORE
+before_install:
+  - sudo rm -f /etc/boto.cfg # see https://github.com/travis-ci/travis-ci/issues/7940 and https://api.travis-ci.org/v3/job/508575291/log.txt
+  - sudo mkdir -p /etc/nix
+  - echo "binary-caches = https://cache.nixos.org/ file://$STORE" | sudo tee -a /etc/nix/nix.conf > /dev/null
+  - echo 'require-sigs = false' | sudo tee -a /etc/nix/nix.conf > /dev/null
+before_cache:
+  - mkdir -p $STORE
+  - nix copy --to file://$STORE -f shell.nix buildInputs
+install:
+  - nix-shell --pure --command "echo 'run nix-shell and exit'"
+script:
+  - nix-shell --pure --command "py.test --doctest-modules"
